requirements.txt
alembic==1.12.1
annotated-types==0.7.0
anyio==3.7.1
asyncpg==0.30.0
bcrypt==4.3.0
bleach==6.2.0
cachetools==5.5.2
certifi==2025.8.3
cffi==1.17.1
charset-normalizer==3.4.3
click==8.1.7
cryptography==45.0.6
Deprecated==1.2.18
distro==1.9.0
ecdsa==0.19.1
fastapi==0.104.1
google-ai-generativelanguage==0.4.0
google-api-core==2.25.1
google-api-python-client==2.179.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-generativeai==0.3.2
googleapis-common-protos==1.70.0
groq==0.4.1
grpcio==1.74.0
grpcio-status==1.62.3
h11==0.16.0
httpcore==1.0.9
httplib2==0.22.0
httptools==0.6.4
httpx==0.25.2
idna==3.10
iniconfig==2.1.0
limits==5.5.0
Mako==1.3.10
MarkupSafe==3.0.2
openai==1.3.7
packaging==25.0
passlib==1.7.4
pluggy==1.6.0
prometheus-client==0.20.0
proto-plus==1.26.1
protobuf==4.25.8
pyasn1==0.6.1
pyasn1_modules==0.4.2
pybreaker==1.4.0
pycparser==2.22
pydantic==2.5.0
pydantic-settings==2.1.0
pydantic_core==2.14.1
pyparsing==3.2.3
pytest==7.4.3
pytest-asyncio==0.21.1
python-dotenv==1.0.0
python-jose==3.3.0
python-multipart==0.0.6
PyYAML==6.0.1
redis==6.4.0
requests==2.31.0
rsa==4.9.1
six==1.17.0
slowapi==0.1.9
sniffio==1.3.1
SQLAlchemy==2.0.23
starlette==0.27.0
structlog==23.2.0
tqdm==4.67.1
typing_extensions==4.15.0
uritemplate==4.2.0
urllib3==2.5.0
uvicorn==0.24.0
uvloop==0.21.0
watchfiles==1.1.0
webencodings==0.5.1
websockets==15.0.1
wrapt==1.17.3
fastapi>=0.104.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
python-multipart>=0.0.6
bleach>=6.0.0
slowapi>=0.1.9
pybreaker>=1.0.1
cryptography>=41.0.0
pydantic[email]>=2.0.0
sqlalchemy>=2.0.0
asyncpg>=0.29.0  # For PostgreSQL
redis>=5.0.0
celery>=5.3.0
